requirements.txt
functions-framework==3.*
Flask==3.1.1
google-genai==1.14.0
pandas==2.2.3
google-cloud-bigquery==3.32.0
numpy==2.2.5
scikit-learn==1.6.1
tenacity==9.1.2
db-dtypes==1.4.3